IDL-compiler: Support for topic keys
Support for topic keys is added. Both '@key' as the '#pragma keylist'
are supported. It is possible to use an inline struct as a key. If
some of its members are defined as keys, only those members actually
are keys. If non of its members are defined as keys, all members are
considered to be keys. The keys are stored with the structure as a
linked list with pointers to the declarations of the struct. The
tests have been extended to cover the topic keys.

The grammar has been adapted such that no spaces are allowed between
identifiers and the '@' and '::' symbols. This is needed to parse
a member declaration such as '@key ::a::s x;' correctly.

Signed-off-by: Frans Faase <frans.faase@adlinktech.com>